Cirrus CI visualization: switch from PNG to SVG
Purpose: SVG includes provisions for highlighting and
de-highlighting regions on mouse hover; I'm using this
to indicate which tasks are disabled on CI:DOC, and
on tags and branches.

This is almost certainly not going to work; I may need to
play with <object> and/or <iframe> or even just give up
entirely.

Signed-off-by: Ed Santiago <santiago@redhat.com>